feat: refactor pagination
